added code to not delete .gitkeep in frames folder

+++ Backend/cleanup.py
@@ -9,5 +9,5 @@ frames_path = 'Frames'
 for filename in os.listdir(frames_path):
     file_path = os.path.join(frames_path, filename)
 
-    if os.path.isfile(file_path):
+    if os.path.isfile(file_path) and filename != '.gitkeep':
         os.remove(file_path)
